:sparkles: feat: 비즈니스 별 예외 분리
