+++ src/main/kotlin/kr/hhplus/be/server/controller/common/exception/ProductException.kt
@@ -1,0 +1,9 @@
+package kr.hhplus.be.server.controller.common.exception
+
+class ProductException(
+    override val message: String,
+    cause: Throwable? = null,
+) : RuntimeException(
+    message,
+    cause,
+)
